Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int x;
+    float x;
     printf("enter the num:");
-    scanf("%d",&x);
+    scanf("%f",&x);
     printf("You entered:");
     return 0;
 }
